api/requirements.txt
fastapi
uvicorn[standard]
pydantic
transformers
torch
chromadb
sentence-transformers
python-multipart
pyyaml
requests
openai>=1.40.0